Add :client_sid concept.
Add a "client_sid" column to the access_tokens table.
The expected usage flow is the following:

1) When requesting an access_token, the client application can
send an additional parameter in the request (named client_sid), which
should contain the session identifier for the client. That is, this value
shoud allow the client to identify/retrieve a session in its application.

2) When a user signs out from the session provider application, if the client
has expiration notifications turned on, the client_sid that was provided during
the token request will be sent back, so that the client can retrieve this session
and do something with it (eg: remove it from the sessions cache).